JOURNAL
This journal contains all of the PowerShell cmdlets that were discussed in the course. It provides thorough explanations, syntax,  examples, and screenshots of every cmdlet in action. The goal is to give users a brief reference guide so they understand and use PowerShell commands with ease. Professionals and students who want to improve their knowledge and proficiency using PowerShell can use this journal as a reference.